Moved contactform into component folder
