+++ .env.example
@@ -1,0 +1,6 @@
+GATSBY_FIREBASE_API_KEY=
+GATSBY_FIREBASE_AUTH_DOMAIN=
+GATSBY_FIREBASE_DATABASE_URL=
+GATSBY_FIREBASE_PROJECT_ID=
+GATSBY_FIREBASE_STORAGE_BUCKET=
+GATSBY_FIREBASE_MESSAGING_SENDER_ID=
