+++ package.json
@@ -11,9 +11,11 @@
     "@fortawesome/react-fontawesome": "^0.1.8",
     "babel-plugin-styled-components": "^1.10.6",
     "bulma": "^0.8.0",
+    "firebase": "^7.6.2",
     "gatsby": "^2.18.12",
     "gatsby-background-image": "^0.9.12",
     "gatsby-image": "^2.2.34",
+    "gatsby-plugin-firebase": "^0.1.7",
     "gatsby-plugin-manifest": "^2.2.31",
     "gatsby-plugin-offline": "^3.0.27",
     "gatsby-plugin-react-helmet": "^3.1.16",
